Commit to switch branches for testing

+++ kernels/ConvNetForward_kernel.cl
@@ -5,14 +5,14 @@
 
 // DO NOT CHANGE THESE DEFAULTS VALUES. IF YOU MUST, MAKE THEM THE SAME LENGTH IN CHARACTERS
 
-double exact_exp(double z) { //from the EXACT CNN builder from Travis Desell
+float exact_exp(float z) { //from the EXACT CNN builder from Travis Desell
     bool is_negative = z < 0;
     if (is_negative) z = -z;
 
     // exp(x) = sum (k = 0 to inf) z^k/k!
-    double result = 1.0 + z;
+    float result = 1.0 + z;
 
-    double prev = z;
+    float prev = z;
     for (short k = 2; k < iterations; k++) {
         prev *= (z / k);
         result += prev;
@@ -26,7 +26,7 @@ double exact_exp(double z) { //from the EXACT CNN builder from Travis Desell
 }
 
 //numthreads should be size of neurons and prevNeurons (should be same)
-__kernel void relu(__global double* prevNeurons, __global double* neurons)
+__kernel void relu(__global float* prevNeurons, __global float* neurons)
 {
 	const int i = get_global_id(0);
 	// neurons[i] = clamp(prevNeurons[i], -RELU_CAP, RELU_CAP);
@@ -39,11 +39,11 @@ __kernel void relu(__global double* prevNeurons, __global double* neurons)
 }
 
 //numthreads should be size of neurons and prevNeurons (should be same)
-__kernel void leakyRelu(__global double* prevNeurons, __global double* neurons)
+__kernel void leakyRelu(__global float* prevNeurons, __global float* neurons)
 {
 	const int i = get_global_id(0);
-	//double newVal = prevNeurons[i] > 0 ? prevNeurons[i] : prevNeurons[i] * .01; 
-	double newVal;
+	//float newVal = prevNeurons[i] > 0 ? prevNeurons[i] : prevNeurons[i] * .01; 
+	float newVal;
 	if(prevNeurons[i] >= 0) 
 		newVal = prevNeurons[i];
 	else 
@@ -61,7 +61,7 @@ __kernel void leakyRelu(__global double* prevNeurons, __global double* neurons)
 
 
 //num threads should be the size of the neurons after the maxPool
-__kernel void maxPool(__global double* prevNeurons, __global double* neurons,
+__kernel void maxPool(__global float* prevNeurons, __global float* neurons,
 	int prevwidth, int prevdepth, int poolsize, int stride)
 {
 	// int width = prevwidth;
@@ -82,7 +82,7 @@ __kernel void maxPool(__global double* prevNeurons, __global double* neurons,
 	i = (i_div_dep/numBlocksPerRow * prevwidth * strxdep + i%prevdepth) + (((i_div_dep)%numBlocksPerRow) * strxdep);
 	
 	int amountToNextLayer = (prevwidth - poolsize) * prevdepth;
-	double maxVal = prevNeurons[i];
+	float maxVal = prevNeurons[i];
 	for(int row = 0; row < poolsize; row++)
 	{
 		for(int col = 0; col < poolsize; col++)
@@ -96,8 +96,8 @@ __kernel void maxPool(__global double* prevNeurons, __global double* neurons,
 	neurons[x] = maxVal;
 }
 
-__kernel void convolve(__global double* prevNeurons, __global double* neurons,
-	__global double* weights, __global double* biases, int numFilters, int filterSize, int stride, int prevwidth, int prevdepth)
+__kernel void convolve(__global float* prevNeurons, __global float* neurons,
+	__global float* weights, __global float* biases, int numFilters, int filterSize, int stride, int prevwidth, int prevdepth)
 {
 	//int myHeight = myBlock/numBlocksPerRow;
 	//int myRowStartIndex = (myBlock/numBlocksPerRow) * width * strxdep;
@@ -124,8 +124,8 @@ __kernel void convolve(__global double* prevNeurons, __global double* neurons,
 
 	//can I do the pointer arithmetic better?
 
-	double result = 0;
-	__global double* curWeight = &(weights[j]);
+	float result = 0;
+	__global float* curWeight = &(weights[j]);
 	for(int a = 0; a < filterSize; a++) //for each layer in the filter
 	{
 		for(int b = 0; b < filterLayerSize; b++)
@@ -141,10 +141,10 @@ __kernel void convolve(__global double* prevNeurons, __global double* neurons,
 	neurons[i] = result + biases[myFilter];
 }
 
-__kernel void convolveConstant(__global double* prevNeurons, __global double* neurons,
-	__constant double* weights, __constant double* biases, int numFilters, int filterSize, int stride, int prevwidth, int prevdepth)
-// __kernel void convolveConstant(__global double* prevNeurons, __global double* neurons,
-// 	__constant double* weights, __constant double* biases, int numFilters, int filterSize, int strxdep, 
+__kernel void convolveConstant(__global float* prevNeurons, __global float* neurons,
+	__constant float* weights, __constant float* biases, int numFilters, int filterSize, int stride, int prevwidth, int prevdepth)
+// __kernel void convolveConstant(__global float* prevNeurons, __global float* neurons,
+// 	__constant float* weights, __constant float* biases, int numFilters, int filterSize, int strxdep, 
 // 	int prevwidth, int amountToNextLayer, int filterLayerSize, int numBlocksPerRow)
 {
 	//int myHeight = myBlock/numBlocksPerRow;
@@ -171,8 +171,8 @@ __kernel void convolveConstant(__global double* prevNeurons, __global double* ne
 
 	//can I do the pointer arithmetic better?
 
-	double result = 0;
-	__constant double* curWeight = &(weights[j]);
+	float result = 0;
+	__constant float* curWeight = &(weights[j]);
 	for(int a = 0; a < filterSize; a++) //for each layer in the filter
 	{
 		for(int b = 0; b < filterLayerSize; b++)
@@ -188,15 +188,15 @@ __kernel void convolveConstant(__global double* prevNeurons, __global double* ne
 	neurons[i] = result + biases[myFilter];
 }
 
-__kernel void softmax(__global double *prevNeurons, __global double *neurons,
-	double denominator)
+__kernel void softmax(__global float *prevNeurons, __global float *neurons,
+	float denominator)
 {
 	int i = get_global_id(0);
 	neurons[i] = exact_exp(prevNeurons[i])/denominator;
 }
 
-__kernel void softmax_allCL(__global double *prevNeurons, __global double *neurons,
-	__global double* denominator)
+__kernel void softmax_allCL(__global float *prevNeurons, __global float *neurons,
+	__global float* denominator)
 {
 	int i = get_global_id(0);
 	neurons[i] = exp(prevNeurons[i])/(*denominator);
@@ -204,7 +204,7 @@ __kernel void softmax_allCL(__global double *prevNeurons, __global double *neuro
 
 }
 
-__kernel void zeroPad(__global double *prevNeurons, __global double *neurons, int pad, int prevwidth,
+__kernel void zeroPad(__global float *prevNeurons, __global float *neurons, int pad, int prevwidth,
 	int prevheight, int depth)
 {
 	int x = get_global_id(0);
@@ -238,12 +238,12 @@ __kernel void zeroPad(__global double *prevNeurons, __global double *neurons, in
 *************************************************/
 
 //make this 2 kernels? one to get max, one to subtract? Prob not worth it without a lot of classes
-__kernel void maxSubtraction(__global double* source, int size)
+__kernel void maxSubtraction(__global float* source, int size)
 {
 	if(size <= 0)
 		return;
-	double max = source[0];
-	double cur;
+	float max = source[0];
+	float cur;
 	for(int i = 1; i < size; i ++)
 	{
 		cur = source[i];
@@ -254,11 +254,11 @@ __kernel void maxSubtraction(__global double* source, int size)
 		source[i] -= max;
 }
 
-__kernel void vectorESum(__global double* source, int size, __global double* denom)
+__kernel void vectorESum(__global float* source, int size, __global float* denom)
 {
 	if(size <= 0)
 		return;
-	double sum = 0;
+	float sum = 0;
 	for(int i=0; i < size; i++)
 		sum += exact_exp(source[i]);
 	*denom = sum;
